html elements are now correctly extending HTMLElement class as per specification
