+++ lib/Element.js
@@ -6,7 +6,6 @@ const { getElementsByClassName, getElementsByTagName } = require("./getElements"
 const Attr = require("./Attr");
 const CSSStyleDeclaration = require("./CSSStyleDeclaration");
 const DOMException = require("domexception");
-const DOMStringMap = require("./DOMStringMap");
 const DOMTokenList = require("./DOMTokenList");
 const HTMLCollection = require("./HTMLCollection");
 const Node = require("./Node");
@@ -15,7 +14,6 @@ const eventnames = require("./eventnames");
 const vm = require("vm");
 
 const classListSymbol = Symbol.for("classList");
-const datasetSymbol = Symbol.for("dataset");
 const rectsSymbol = Symbol.for("rects");
 const scrolledSymbol = Symbol.for("scrolled");
 const stylesSymbol = Symbol.for("styles");
@@ -37,7 +35,6 @@ module.exports = class Element extends Node {
 
     this[scrolledSymbol] = {left: 0, top: 0};
     this[classListSymbol] = new DOMTokenList(this);
-    this[datasetSymbol] = new DOMStringMap(this);
     this[stylesSymbol] = new CSSStyleDeclaration(this);
 
     if ($elm[0].attribs) {
@@ -204,9 +201,6 @@ module.exports = class Element extends Node {
   }
   get offsetHeight() {
     return this[rectsSymbol].height;
-  }
-  get dataset() {
-    return this[datasetSymbol];
   }
   get style() {
     return this[stylesSymbol];
