+++ lib/HTMLAnchorElement.js
@@ -1,10 +1,10 @@
 "use strict";
 
-const Element = require("./Element");
+const HTMLElement = require("./HTMLElement");
 
 const urlSymbol = Symbol.for("href");
 
-module.exports = class HTMLAnchorElement extends Element {
+module.exports = class HTMLAnchorElement extends HTMLElement {
   get href() {
     return hrefPropertyGetter(this[urlSymbol], "href");
   }
