+++ lib/HTMLButtonElement.js
@@ -1,9 +1,9 @@
 "use strict";
 
 const { PointerEvent } = require("./Events");
-const Element = require("./Element");
+const HTMLElement = require("./HTMLElement");
 
-module.exports = class HTMLButtonElement extends Element {
+module.exports = class HTMLButtonElement extends HTMLElement {
   constructor(document, $elm) {
     super(document, $elm);
     this.ownerDocument.addEventListener("click", (e) => {
