+++ lib/HTMLDialogElement.js
@@ -1,9 +1,9 @@
 "use strict";
 
 const { Event } = require("./Events");
-const Element = require("./Element");
+const HTMLElement = require("./HTMLElement");
 
-module.exports = class HTMLDialogElement extends Element {
+module.exports = class HTMLDialogElement extends HTMLElement {
   constructor(document, $elm) {
     super(document, $elm);
