+++ lib/HTMLElement.js
@@ -1,5 +1,22 @@
 "use strict";
 
 const Element = require("./Element");
+const DOMStringMap = require("./DOMStringMap");
 
-module.exports = class HTMLElement extends Element {};
+const datasetSymbol = Symbol.for("dataset");
+
+module.exports = class HTMLElement extends Element {
+  constructor(document, $elm) {
+    super(document, $elm);
+
+    this[datasetSymbol] = new DOMStringMap(this);
+  }
+
+  get dataset() {
+    return this[datasetSymbol];
+  }
+
+  get contentEditable() {
+    return this.hasAttribute("contenteditable") || false;
+  }
+};
