+++ lib/HTMLFormElement.js
@@ -1,13 +1,13 @@
 "use strict";
 
 const { Event, SubmitEvent, symbols } = require("./Events");
-const Element = require("./Element");
+const HTMLElement = require("./HTMLElement");
 const HTMLFormControlsCollection = require("./HTMLFormControlsCollection");
 const HTMLInputElement = require("./HTMLInputElement");
 
 const originSymbol = Symbol.for("origin");
 
-module.exports = class HTMLFormElement extends Element {
+module.exports = class HTMLFormElement extends HTMLElement {
   constructor(document, $elm) {
     super(document, $elm);
