+++ lib/HTMLImageElement.js
@@ -1,8 +1,8 @@
 "use strict";
 
-const Element = require("./Element");
+const HTMLElement = require("./HTMLElement");
 
-module.exports = class HTMLImageElement extends Element {
+module.exports = class HTMLImageElement extends HTMLElement {
   constructor(document, $elm) {
     super(document, $elm);
   }
