+++ lib/HTMLInputElement.js
@@ -1,13 +1,13 @@
 "use strict";
 
 const { Event, InputEvent, PointerEvent } = require("./Events");
-const Element = require("./Element");
+const HTMLElement = require("./HTMLElement");
 const ValidityState = require("./ValidityState");
 
 const kValidity = Symbol.for("validity");
 const kValidationMessage = Symbol.for("validation message");
 
-module.exports = class HTMLInputElement extends Element {
+module.exports = class HTMLInputElement extends HTMLElement {
   constructor(...args) {
     super(...args);
