+++ lib/HTMLOptionElement.js
@@ -1,9 +1,9 @@
 "use strict";
 
-const Element = require("./Element");
+const HTMLElement = require("./HTMLElement");
 const {Event} = require("./Events");
 
-module.exports = class HTMLOptionElement extends Element {
+module.exports = class HTMLOptionElement extends HTMLElement {
   get selected() {
     return this.$elm.prop("selected");
   }
