+++ lib/HTMLScriptElement.js
@@ -1,9 +1,9 @@
 "use strict";
 
-const Element = require("./Element");
+const HTMLElement = require("./HTMLElement");
 const vm = require("vm");
 
-module.exports = class HTMLScriptElement extends Element {
+module.exports = class HTMLScriptElement extends HTMLElement {
   get textContent() {
     return this.$elm.html();
   }
