+++ lib/HTMLSelectElement.js
@@ -1,14 +1,14 @@
 "use strict";
 
 const { Event } = require("./Events");
-const Element = require("./Element");
+const HTMLElement = require("./HTMLElement");
 const HTMLCollection = require("./HTMLCollection");
 const HtmlOptionCollection = require("./HTMLOptionsCollection");
 
 const kOptions = Symbol.for("options");
 const kSelectedOptions = Symbol.for("selected options");
 
-module.exports = class HTMLSelectElement extends Element {
+module.exports = class HTMLSelectElement extends HTMLElement {
   constructor(...args) {
     super(...args);
     this[kOptions] = new HtmlOptionCollection(this);
