+++ lib/HTMLTemplateElement.js
@@ -1,8 +1,8 @@
 "use strict";
 
-const Element = require("./Element");
+const HTMLElement = require("./HTMLElement");
 
-module.exports = class HTMLTemplateElement extends Element {
+module.exports = class HTMLTemplateElement extends HTMLElement {
   get content() {
     const DocumentFragment = require("./DocumentFragment");
     return new DocumentFragment(this.ownerDocument, this);
