+++ lib/HTMLTextAreaElement.js
@@ -1,9 +1,9 @@
 "use strict";
 
 const { InputEvent } = require("./Events");
-const Element = require("./Element");
+const HTMLElement = require("./HTMLElement");
 
-module.exports = class HTMLTextAreaElement extends Element {
+module.exports = class HTMLTextAreaElement extends HTMLElement {
   get disabled() {
     return this.$elm.prop("disabled");
   }
