+++ lib/HTMLVideoElement.js
@@ -1,8 +1,8 @@
 "use strict";
 
-const Element = require("./Element");
+const HTMLElement = require("./HTMLElement");
 
-module.exports = class HTMLVideoElement extends Element {
+module.exports = class HTMLVideoElement extends HTMLElement {
   constructor(document, $elm) {
     super(document, $elm);
   }
