+++ lib/elementFactory.js
@@ -1,7 +1,7 @@
 "use strict";
 
 const DocumentType = require("./DocumentType");
-const Element = require("./Element");
+const HTMLElement = require("./HTMLElement");
 const HTMLAnchorElement = require("./HTMLAnchorElement");
 const HTMLButtonElement = require("./HTMLButtonElement");
 const HTMLDialogElement = require("./HTMLDialogElement");
@@ -55,7 +55,7 @@ ElementFactory.prototype.create = function define($elm) {
     return elm;
   }
 
-  return new Element(document, $elm);
+  return new HTMLElement(document, $elm);
 };
 
 ElementFactory.prototype.define = function define(tagName, TagClass) {
